backend/app/requirements.txt
django==6.0.4
psycopg[binary]==3.3.3
djangorestframework==3.17.1
stripe